requirements.txt
evds
pandas>=2.0.0
SQLAlchemy>=2.0.0
pymssql>=2.3.0
python-dotenv>=1.0.0
tzdata>=2023.3
